+++ Userland/SampleCodeModule/tests/test_processes.c
@@ -1,3 +1,5 @@
+// This is a personal academic project. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++ and C#: http://www.viva64.com
 #include "lib.h"
 #include "test_util.h"
 
@@ -24,7 +26,7 @@ int64_t test_processes(uint64_t argc, char *argv[]) {
     return -1;
 
   printf("=== TEST DE PROCESOS INICIADO ===\n");
-  printf("Numero maximo de procesos: %d\n", max_processes);
+  printf("Numero maximo de procesos: %llu\n", (unsigned long long)max_processes);
 
   p_rq p_rqs[max_processes];
